requirements.txt
# Core dependencies
aiohttp>=3.9.0
python-dotenv>=1.0.0
elasticsearch==8.12.1
PyYAML>=6.0

# Web UI
streamlit>=1.28.0

# API and HTTP
requests>=2.31.0
urllib3>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
flake8>=6.1.0
black>=23.12.0
isort>=5.13.0
mypy>=1.7.0

# Monitoring and logging
python-json-logger>=2.0.7

# Security
cryptography>=41.0.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0
